Modularize SofaGeneralVisual
