+++ SofaGeneral/CMakeLists.txt
@@ -17,7 +17,6 @@ set(SOFAGENERAL_TARGETS
     SofaGeneralRigid
     SofaGeneralSimpleFem
     SofaGeneralTopology
-    SofaGeneralVisual
     SofaBoundaryCondition
     SofaConstraint
     SofaGeneralEngine
@@ -28,6 +27,14 @@ set(SOFAGENERAL_TARGETS
 
 foreach(TARGET ${SOFAGENERAL_TARGETS})
     add_subdirectory(../modules/${TARGET} ${CMAKE_CURRENT_BINARY_DIR}/${TARGET})
+endforeach()
+
+# Modules
+set(SOFAGENERAL_MODULES
+    SofaGeneralVisual
+)
+foreach(MODULE ${SOFAGENERAL_MODULES})
+    find_package(${MODULE} REQUIRED)
 endforeach()
 
 set(SOFAGENERAL_SRC src/${PROJECT_NAME})
@@ -42,7 +49,7 @@ set(SOURCE_FILES
     )
 
 add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES})
-target_link_libraries(${PROJECT_NAME} PUBLIC ${SOFAGENERAL_TARGETS})
+target_link_libraries(${PROJECT_NAME} PUBLIC ${SOFAGENERAL_TARGETS} ${SOFAGENERAL_MODULES})
 
 sofa_create_package_with_targets(
     PACKAGE_NAME ${PROJECT_NAME}
