+++ examples/Components/visualmodel/NavigationRecordedCamera.scn
@@ -1,6 +1,8 @@
 <?xml version="1.0"?>
 <Node 	 name="root"  gravity="0 0 -9.81"  dt="0.01" >
-    <RequiredPlugin name="SofaOpenglVisual"/>
+  <RequiredPlugin name="SofaOpenglVisual"/>
+  <RequiredPlugin name="SofaGeneralVisual"/>
+    
   <VisualStyle name="visualStyle1"  displayFlags="showBehavior" />
   <MeshObjLoader name="meshLoader_0" filename="mesh/dragon.obj" handleSeams="1" />
   <OglModel template="ExtVec3f" name="Objective" src="@meshLoader_0"  primitiveType="DEFAULT"  blendEquation="GL_FUNC_ADD"  sfactor="GL_SRC_ALPHA"  dfactor="GL_ONE_MINUS_SRC_ALPHA" />
