+++ examples/Components/visualmodel/RecordedCamera.scn
@@ -1,6 +1,8 @@
 <!-- For more details see: https://wiki.sofa-framework.org/tdev/wiki/Notes/NewLoaderArchitecture -->
 <Node name="Root" gravity="0 -9.81 0" dt="0.15">
     <RequiredPlugin name="SofaOpenglVisual"/>
+    <RequiredPlugin name="SofaGeneralVisual"/>
+
     <VisualStyle displayFlags="showVisual" />
     <DefaultPipeline name="DefaultCollisionPipeline" verbose="0" draw="0" depth="6" />
     <BruteForceDetection name="Detection" />
