+++ modules/SofaGeneralVisual/CMakeLists.txt
@@ -1,30 +1,45 @@
 cmake_minimum_required(VERSION 3.12)
-project(SofaGeneralVisual)
+project(SofaGeneralVisual VERSION 1.0 LANGUAGES CXX)
+
+find_package(SofaBase REQUIRED)
+
+set(SOFAGENERALVISUAL_SRC "src/${PROJECT_NAME}")
 
 set(HEADER_FILES
-    config.h
-    initGeneralVisual.h
+    ${SOFAGENERALVISUAL_SRC}/config.h.in
+    ${SOFAGENERALVISUAL_SRC}/initSofaGeneralVisual.h
     )
 
 set(SOURCE_FILES
-    initGeneralVisual.cpp
+    ${SOFAGENERALVISUAL_SRC}/initSofaGeneralVisual.cpp
     )
 
 list(APPEND HEADER_FILES
-    RecordedCamera.h
-    VisualTransform.h
-    Visual3DText.h
+    ${SOFAGENERALVISUAL_SRC}/RecordedCamera.h
+    ${SOFAGENERALVISUAL_SRC}/VisualTransform.h
+    ${SOFAGENERALVISUAL_SRC}/Visual3DText.h
     )
 list(APPEND SOURCE_FILES
-    RecordedCamera.cpp
-    VisualTransform.cpp
-    Visual3DText.cpp
+    ${SOFAGENERALVISUAL_SRC}/RecordedCamera.cpp
+    ${SOFAGENERALVISUAL_SRC}/VisualTransform.cpp
+    ${SOFAGENERALVISUAL_SRC}/Visual3DText.cpp
     )
 
 add_library(${PROJECT_NAME} SHARED ${HEADER_FILES} ${SOURCE_FILES})
 target_link_libraries(${PROJECT_NAME} PUBLIC SofaBaseTopology SofaBaseVisual)
 
-sofa_add_targets_to_package(
-    PACKAGE_NAME SofaGeneral
+sofa_create_package_with_targets(
+    PACKAGE_NAME ${PROJECT_NAME}
+    PACKAGE_VERSION ${PROJECT_VERSION}
     TARGETS ${PROJECT_NAME} AUTO_SET_TARGET_PROPERTIES
-    )
+    INCLUDE_SOURCE_DIR "src"
+    INCLUDE_INSTALL_DIR "${PROJECT_NAME}"
+)
+
+# Tests
+# If SOFA_BUILD_TESTS exists and is OFF, then these tests will be auto-disabled
+# cmake_dependent_option(SOFAGENERALVISUAL_BUILD_TESTS "Compile the automatic tests" ON "SOFA_BUILD_TESTS OR NOT DEFINED SOFA_BUILD_TESTS" OFF)
+# if(SOFAGENERALVISUAL_BUILD_TESTS)
+#     enable_testing()
+#     add_subdirectory(${PROJECT_NAME}_test)
+# endif()
