+++ modules/SofaGeneralVisual/README.md
@@ -1,0 +1,1 @@
+# SofaGeneralVisual
